Rewrote some stuff
Still very confused as to why the .accept method continues to create
more than one thread for the same client over and over. There has to be
some verification that we aren't connecting the same client multiple
times. Maybe a list containing currently connected ips would do it.
Otherwise it might be a quick fix I'm not familiar with. I'll keep
digging.