Some more changes
Made the mistake of attempting to create a persistent tcp connection. In
other words it's fine for something like Firefox to send multiple
requests since every thread created doesn't have to return something
back to the client.
Now it's just a matter of interpreting queries created by the client and
returning the appropriate response.