Major rewrites and a big socket issue
Alright, so here's the deal. I have finally been able to get the url
token sent to the internet, and it also returns the html in a nice
String variable. However, I can't seem to write back to the connected
browser (This is where we use the clientSocket.getOutputStream to send
stuff). Something during the writeback process causes the socket to
close. I have no idea what it is. Please take a look at it when you can,
but I'll keep working on it.